in progress of switching bufferDo to free
